requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aioprocessing==2.0.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
bitarray==2.7.6
ccxt==4.0.16
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.1
cytoolz==0.12.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-retry==0.1.18
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
flashbots==1.1.1
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
influxdb-client==1.36.1
ipfshttpclient==0.8.0a2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
lru-dict==1.2.0
multiaddr==0.0.9
multicall==0.7.4
multidict==6.0.4
nest-asyncio==1.5.6
netaddr==0.8.0
numpy==1.25.0
parsimonious==0.8.1
protobuf==3.19.5
pycares==4.3.0
pycparser==2.21
pycryptodome==3.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
reactivex==4.0.4
referencing==0.29.1
requests==2.31.0
rlp==2.0.1
rpds-py==0.8.10
six==1.16.0
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.3
varint==1.0.2
web3==5.31.4
websockets==9.1
yarl==1.9.2
